requirements.txt
autopep8==1.5.2
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
